+++ test/specs/mainnet/ln.e2e.ts
@@ -6,26 +6,12 @@ import {
   restoreWallet,
   tap,
   sleep,
-  expectTextWithin,
-  doNavigationClose,
 } from '../../helpers/actions';
+import { waitForMainnetWalletReady } from '../../helpers/mainnet';
 import { ciIt } from '../../helpers/suite';
 
 const PAYMENT_TIMEOUT_MS = 90_000;
-const WALLET_SYNC_TIMEOUT_MS = 90_000;
-const APP_STATUS_ROW_TIMEOUT_MS = 90_000;
 const SCREEN_TRANSITION_TIMEOUT_MS = 30_000;
-
-function resolveLnStabilizeDelayMs(): number {
-  const fromEnv = process.env.LN_STABILIZE_DELAY_MS;
-  if (fromEnv) {
-    const parsed = Number.parseInt(fromEnv, 10);
-    if (Number.isFinite(parsed) && parsed >= 0) {
-      return parsed;
-    }
-  }
-  return process.env.CI ? 45_000 : 10_000;
-}
 
 const ERROR_TOASTS = ['PaymentFailedToast', 'ExpiredLightningToast', 'InsufficientSpendingToast'];
 
@@ -70,29 +56,6 @@ function resolveMainnetLnReceiver(config: MainnetLnSuiteConfig): MainnetLnReceiv
     lnAddress: resolveEnvValue(config.addressEnv),
     amountSats: resolveAmountSats(config.amountEnv, config.defaultAmountSats),
   };
-}
-
-async function waitForWalletReady(): Promise<void> {
-  console.info('→ [LN] Waiting for wallet home screen...');
-  await elementById('TotalBalance-primary').waitForDisplayed({ timeout: WALLET_SYNC_TIMEOUT_MS });
-  const stabilizeMs = resolveLnStabilizeDelayMs();
-  console.info(`→ [LN] Home screen ready, letting LN node stabilize (${stabilizeMs / 1000}s)...`);
-  await sleep(stabilizeMs);
-  console.info('→ [LN] Verify app status is ready');
-  await tap('HeaderMenu');
-  await tap('DrawerAppStatus');
-
-  await expectTextWithin('Status-internet', 'Connected', { timeout: APP_STATUS_ROW_TIMEOUT_MS });
-  await expectTextWithin('Status-electrum', 'Connected', { timeout: APP_STATUS_ROW_TIMEOUT_MS });
-  await expectTextWithin('Status-lightning_node', 'Running', {
-    timeout: APP_STATUS_ROW_TIMEOUT_MS,
-  });
-  await expectTextWithin('Status-lightning_connection', 'Open', {
-    timeout: APP_STATUS_ROW_TIMEOUT_MS,
-  });
-
-  await doNavigationClose();
-  console.info('→ [LN] App status verified');
 }
 
 async function waitForAmountScreen(): Promise<void> {
@@ -145,7 +108,7 @@ async function sendPaymentToLnAddress(receiver: MainnetLnReceiver): Promise<void
     expectAndroidAlert: false,
   });
 
-  await waitForWalletReady();
+  await waitForMainnetWalletReady({ logPrefix: 'LN' });
 
   console.info(`→ [LN] Entering address: ${receiver.lnAddress}`);
   await enterAddress(receiver.lnAddress, { acceptCameraPermission: false, addressTimeout: 60_000 });
